Static link Windows.
Windows users deal with distributing OpenAL DLLs, which is annoying.
Specifying the .a library forces compilers to compile statically.
Because the .dll.a library has some extra symbols that the .a doesn't,
those symbols need to be added back by linking dynamically to libstdc++,
libole32, and libwinmm. These are standard Windows DLLs,
and do not need to be distributed.